final_pipe/airflow/requirements.txt
apache-airflow==2.8.3
apache-airflow-providers-google==10.16.0
joblib~=1.0
numpy~=1.20
scikit-learn~=0.24
google-cloud-storage>=1.26.0,<2.0.0dev
google-cloud-aiplatform[prediction]>=1.16.0
pandas
pandas-gbq